feat(tests): add fuzz tests for NormalizeToolInput and MergeExtra functions
