+++ ai/content_test.go
@@ -83,3 +83,22 @@ func TestNormalizeToolInput(t *testing.T) {
 		})
 	}
 }
+
+func FuzzNormalizeToolInput(f *testing.F) {
+	f.Add([]byte(""))
+	f.Add([]byte("null"))
+	f.Add([]byte("  {}  "))
+	f.Add([]byte(`{"a":1}`))
+	f.Add([]byte("[1,2,3]"))
+	f.Add([]byte("not json"))
+	f.Fuzz(func(t *testing.T, in []byte) {
+		out := ai.NormalizeToolInput(json.RawMessage(in))
+		// Postcondition: output is never empty and never the JSON null literal.
+		if len(out) == 0 {
+			t.Fatalf("normalized output is empty for input %q", in)
+		}
+		if string(out) == "null" {
+			t.Fatalf("normalized output is null for input %q", in)
+		}
+	})
+}
